+++ src/containers/App/App.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Router from "../Router";
+
+function App() {
+  return (
+    <>
+      <Router />
+    </>
+  );
+}
+
+export default App;
