+++ src/containers/Home/index.tsx
@@ -1,0 +1,19 @@
+import logo from '../../logo.svg';
+import React from "react";
+
+import "./Home.scss";
+
+function Home() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Welcome to Faba onboarding project - Job board
+        </p>
+      </header>
+    </div>
+  );
+}
+
+export default Home;
