+++ src/containers/Router/services/index.ts
@@ -1,0 +1,3 @@
+export const routes = {
+  ROOT: "/",
+};
